modification de la table chambre, prix en float noa

+++ src/Entity/Chambre.php
@@ -20,7 +20,7 @@ class Chambre
     private ?string $type = null;
 
     #[ORM\Column]
-    private ?int $prix = null;
+    private ?float $prix = null;
 
     public function getId(): ?int
     {
@@ -51,12 +51,12 @@ class Chambre
         return $this;
     }
 
-    public function getPrix(): ?int
+    public function getPrix(): ?float
     {
         return $this->prix;
     }
 
-    public function setPrix(int $prix): static
+    public function setPrix(float $prix): static
     {
         $this->prix = $prix;
